backend/pyproject.toml
[project]
name = "rag-chatbot-backend"
version = "1.0.0"
description = "RAG Chatbot Backend Integration with OpenAI Agent Controller"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "cohere>=5.20.0",
    "fastapi>=0.115.6",
    "openai>=1.57.4",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "requests>=2.32.5",
    "uvicorn>=0.36.2",
    "pydantic>=2.10.3",
    "httpx>=0.28.0"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools"
